fix: pick the most dpi-like where a country has more than one
When filtering the dpi data down to 1 system per coutnry, ensure we pick the most dpi-like system where there is more than 1 to choose from.

Previously we assumed the data would always list the more dpi-like system before the less dpi-like, but this may not be the case.

fixes: https://github.com/olizilla/digital-public-infra-map/issues/29

License: MIT
Signed-off-by: Oli Evans <oli@zilla.org.uk>