fix: pick the most dpi-like where a country has more than one (#30)
When filtering the dpi data down to 1 system per coutnry, ensure we pick
the most dpi-like system where there is more than 1 to choose from.

Previously we assumed the data would always list the more dpi-like
system before the less dpi-like, but this may not be the case.

fixes: https://github.com/olizilla/digital-public-infra-map/issues/29

Note that this bug was not the cause of the difference in the headline
stats from the spreadsheet calculation. The website is currently
displaying the count of "DPI-like systems **per country**" so where a
country has more than 1 DPI-like system for a given category, it only
counts as 1 for the headline stat. I believe the spreadsheet calculation
is counting "number of DPI-like systems".

License: MIT

Signed-off-by: Oli Evans <oli@zilla.org.uk>