feat: test out stale-while-revalidate in _headers file (#41)
Add a `_headers` file to add `Cache-Control` headers with
stale-while-revalidate to html responses.

The intention is to ratchet these up as we don't redeploy the site that
often, the map html is quite large, and ideally we can make the site
work offline for the duration of the stale-while-revalidate period.

see: https://developers.cloudflare.com/pages/configuration/headers/

License: MIT

---------

Signed-off-by: Oli Evans <oli@zilla.org.uk>